Ensure static linking to libssl and libcrypto
Otherwise installers give out a load of ugly warnings until
the point when OpenSSL and libgcc-ng have gotten installed.

I tried installing in dependency order (and intend to keep
that enabled) but it does not help. Python always gets
installed first (at present). Even if it did not, we would
still see warnings for OpenSSL and libgcc-ng.

Interesingly, the fully-static Python 3.6 does not suffer
from this so that could be an option if this does not work.