Drop adding Library/bin to PATH on Windows.
Ping @msarahan, should we do this? I feel like if we want to go down the
route of managing this stuff oversevles (which this latest patch does
then, if we did want to ensure the dirs were on the DLL search path we'd
do it at a lower level via AddDllDirectory() rather than adding it to
PATH (which should still work given the latest changes by Kai so we can
re-call the function).

Python 3.7 hasn't done this for a while and hopefully we have a good
grasp on all the issues now.